+++ packages/jest-validate/src/__tests__/validateCLIOptions.test.js
